more complete typer example and split to 3rd party compat
